feat: unify StickyCTA implementation and update mobile-header.css for hide rule
