+++ src/index.css
@@ -1,27 +1,32 @@
-/* Import must be first (before any @tailwind or rules) */
-@import './styles/accents.css';
-
-/* Tailwind layers */
+/* Tailwind layers (must come before any custom imports/styles for correct layering) */
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
 
-/* Global typography: prefer Inter or system sans for clean, readable UI */
-:root {
-  --color-primary: #005ea2;       /* navy */
-  --color-primary-hover: #0078c8;
-  --color-accent: #1dd380;        /* teal/green */
-}
-
-html {
-  font-family: Inter, ui-sans-serif, system-ui, -apple-system, Segoe UI, Roboto,
-    Helvetica, Arial, "Apple Color Emoji", "Segoe UI Emoji";
-}
-
-/* Utility container used in header */
-.container-default {
-  @apply mx-auto max-w-7xl px-4 sm:px-6 lg:px-8;
-}
-
-
+/* Custom imports */
+@import "./styles/accents.css";
 @import "./styles/mobile-header.css";
+
+/* Base layer: tokens + typography */
+@layer base {
+  :root {
+    --color-primary: #005ea2;       /* navy */
+    --color-primary-hover: #0078c8;
+    --color-accent: #1dd380;        /* teal/green */
+  }
+
+  html {
+    font-family: Inter, ui-sans-serif, system-ui, -apple-system, Segoe UI, Roboto,
+      Helvetica, Arial, "Apple Color Emoji", "Segoe UI Emoji";
+  }
+}
+
+/* Components layer */
+@layer components {
+  /* Utility container used in header */
+  .container-default {
+    @apply mx-auto max-w-7xl px-4 sm:px-6 lg:px-8;
+  }
+}
+
+
